requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==46.0.5
    # via pyjwt
duckdb==1.4.3
    # via mcf (pyproject.toml)
fastapi==0.128.0
    # via mcf (pyproject.toml)
filelock==3.20.3
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2026.1.0
    # via
    #   huggingface-hub
    #   torch
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via mcf (pyproject.toml)
huggingface-hub==0.36.0
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via torch
joblib==1.5.3
    # via scikit-learn
lxml==6.0.2
    # via python-docx
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.6.1
    # via torch
numpy==2.4.0
    # via
    #   mcf (pyproject.toml)
    #   scikit-learn
    #   scipy
    #   transformers
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
psycopg2-binary==2.9.11
    # via mcf (pyproject.toml)
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   mcf (pyproject.toml)
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via mcf (pyproject.toml)
pygments==2.19.2
    # via rich
pyjwt==2.11.0
    # via mcf (pyproject.toml)
pypdf==6.7.5
    # via mcf (pyproject.toml)
python-docx==1.2.0
    # via mcf (pyproject.toml)
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.21
    # via mcf (pyproject.toml)
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2026.1.15
    # via transformers
requests==2.32.5
    # via
    #   huggingface-hub
    #   transformers
rich==14.2.0
    # via
    #   mcf (pyproject.toml)
    #   typer
safetensors==0.7.0
    # via transformers
scikit-learn==1.8.0
    # via sentence-transformers
scipy==1.17.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==5.2.0
    # via mcf (pyproject.toml)
setuptools==80.9.0
    # via torch
shellingham==1.5.4
    # via typer
starlette==0.50.0
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via mcf (pyproject.toml)
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.22.2
    # via transformers
torch==2.9.1
    # via sentence-transformers
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   sentence-transformers
    #   transformers
transformers==4.57.5
    # via sentence-transformers
typer==0.21.0
    # via mcf (pyproject.toml)
typing-extensions==4.15.0
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sentence-transformers
    #   torch
    #   typer
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn==0.40.0
    # via mcf (pyproject.toml)
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
